Convert announce.gradle to Kotlin
First actual Gradle module in the list.

Introduce Kotlin extensions in buildSrc for type safe access to:
- libraries defined in gradle/dependencies.gradle
- useTestFixture() Groovy Closure as Project extra property

We should strive to make this type safety less artificial, and better.
But the fact that dependencies.gradle is both applied to the build and
its own buildSrc would require a separate shared compilation unit. This
is good enough for now.

Signed-off-by: Paul Merlin <paul@gradle.com>